refactor: extract racer-stats.ts, DenPlacementStrip, DEN_LIGHT
- Move bestLane + computeRacerStats to lib/racer-stats.ts (re-exported
  from certificate-stats.ts for backwards compat)
- Extract DenPlacementStrip component from IIFE in RacerProfileView
- Consolidate light-accent den logic into DEN_LIGHT set in den-utils.ts
  (replaces TOO_LOW_CONTRAST_DENS and inline isLight arrays)
